+++ src/main/java/com/shinhan/knockknock/controller/CardHistoryController.java
@@ -5,8 +5,8 @@ import com.shinhan.knockknock.domain.dto.cardhistory.CreateCardHistoryRequest;
 import com.shinhan.knockknock.domain.dto.cardhistory.ReadCardHistoryResponse;
 import com.shinhan.knockknock.domain.entity.CardEntity;
 import com.shinhan.knockknock.repository.CardRepository;
-import com.shinhan.knockknock.service.CardCategoryService;
-import com.shinhan.knockknock.service.CardHistoryService;
+import com.shinhan.knockknock.service.cardcategory.CardCategoryService;
+import com.shinhan.knockknock.service.cardhistory.CardHistoryService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
